fix collage generation for short files
PIL thumbnail() does not enlarge images.
For a short video clip there might not be enough extracted frames to fill the width of the collage base,
while keeping the near-square shape, without enlarging the images.
This results in huge (black) gaps between the individual frames on the collage, ruining the image processing afterwards.

This could've been solved in a couple of different ways.
This was the lazy approach where I don't have to change anything else.